requirements.txt
feedparser>=6.0.11
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
schedule==1.2.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.11
gunicorn==21.2.0
Flask==3.0.0
